requirements.txt
langchain
langchain-community
langchain-openai
chromadb
langchain-chroma
tiktoken
python-dotenv
pypdf
requests
beautifulsoup4
types-beautifulsoup4